Events refactored. Corresponding documentation improved.
